refactor: move `shouldRunCheck` back
