+++ .changeset/old-bees-study.md
@@ -1,5 +1,4 @@
 ---
-"@pnpm/deps.should-run-check": major
 "@pnpm/plugin-commands-script-runners": patch
 "@pnpm/lifecycle": patch
 "pnpm": patch
