--- deps/should-run-check/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@pnpm/deps.should-run-check",
-  "version": "1000.0.0",
-  "description": "Determine whether a deps check can run",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib",
-    "!*.map"
-  ],
-  "engines": {
-    "node": ">=18.12"
-  },
-  "scripts": {
-    "lint": "eslint \"src/**/*.ts\" \"test/**/*.ts\"",
-    "_test": "jest",
-    "test": "pnpm run compile && pnpm run _test",
-    "prepublishOnly": "pnpm run compile",
-    "fix": "tslint -c tslint.json src/**/*.ts test/**/*.ts --fix",
-    "compile": "tsc --build && pnpm run lint --fix"
-  },
-  "repository": "https://github.com/pnpm/pnpm/blob/main/deps/should-run-check",
-  "keywords": [
-    "pnpm10",
-    "pnpm"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/pnpm/pnpm/issues"
-  },
-  "homepage": "https://github.com/pnpm/pnpm/blob/main/deps/should-run-check#readme",
-  "devDependencies": {
-    "@pnpm/deps.should-run-check": "workspace:*"
-  },
-  "funding": "https://opencollective.com/pnpm",
-  "exports": {
-    ".": "./lib/index.js"
-  },
-  "jest": {
-    "preset": "@pnpm/jest-config"
-  }
-}
