--- deps/should-run-check/test/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "../test.lib",
-    "rootDir": "."
-  },
-  "include": [
-    "**/*.ts",
-    "../../../__typings__/**/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": ".."
-    }
-  ]
-}
