--- deps/should-run-check/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@pnpm/tsconfig",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "src"
-  },
-  "include": [
-    "src/**/*.ts",
-    "../../__typings__/**/*.d.ts"
-  ],
-  "references": []
-}
