--- deps/should-run-check/tsconfig.lint.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/**/*.ts",
-    "test/**/*.ts",
-    "../../__typings__/**/*.d.ts"
-  ]
-}
