+++ exec/plugin-commands-script-runners/src/run.ts
@@ -10,7 +10,6 @@ import { prepareExecutionEnv } from '@pnpm/plugin-commands-env'
 import { FILTERING, UNIVERSAL_OPTIONS } from '@pnpm/common-cli-options-help'
 import { type Config, types as allTypes } from '@pnpm/config'
 import { PnpmError } from '@pnpm/error'
-import { DISABLE_DEPS_CHECK_ENV, shouldRunCheck } from '@pnpm/deps.should-run-check'
 import { type CheckDepsStatusOptions } from '@pnpm/deps.status'
 import {
   runLifecycleHook,
@@ -26,6 +25,7 @@ import { existsInDir } from './existsInDir'
 import { handler as exec } from './exec'
 import { buildCommandNotFoundHint } from './buildCommandNotFoundHint'
 import { runDepsStatusCheck } from './runDepsStatusCheck'
+import { DISABLE_DEPS_CHECK_ENV, shouldRunCheck } from './shouldRunCheck'
 
 export const IF_PRESENT_OPTION: Record<string, unknown> = {
   'if-present': Boolean,
