+++ exec/plugin-commands-script-runners/src/shouldRunCheck.ts
