+++ exec/plugin-commands-script-runners/test/shouldRunCheck.test.ts
@@ -1,4 +1,4 @@
-import { DISABLE_DEPS_CHECK_ENV, shouldRunCheck } from '../src/index'
+import { DISABLE_DEPS_CHECK_ENV, shouldRunCheck } from '../src/shouldRunCheck'
 
 test('should return true if skip env is not defined and script name is not special', () => {
   expect(shouldRunCheck({}, 'start')).toBe(true)
