+++ exec/plugin-commands-script-runners/tsconfig.json
@@ -34,9 +34,6 @@
       "path": "../../crypto/hash"
     },
     {
-      "path": "../../deps/should-run-check"
-    },
-    {
       "path": "../../deps/status"
     },
     {
