+++ pnpm/package.json
@@ -36,7 +36,6 @@
     "@pnpm/crypto.hash": "workspace:*",
     "@pnpm/default-reporter": "workspace:*",
     "@pnpm/dependency-path": "workspace:*",
-    "@pnpm/deps.should-run-check": "workspace:*",
     "@pnpm/env.path": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/filter-workspace-packages": "workspace:*",
