+++ pnpm/tsconfig.json
@@ -61,9 +61,6 @@
       "path": "../crypto/hash"
     },
     {
-      "path": "../deps/should-run-check"
-    },
-    {
       "path": "../env/path"
     },
     {
